@@ -1,76 +1,59 @@
 {
-  "name": "fin-bot-api",
+  "name": "finbots-front",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
-    "test:ci": "ng test --karmaConfig=src/karma-ci.conf.js --codeCoverage=true",
-    "precommit": "run-s format:fix lint-staged",
-    "lint-staged": "lint-staged",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
     "format:fix": "pretty-quick --staged",
     "format:check": "prettier --config ./.prettierrc.json --list-different \"src/{app,environments,assets}/**/*{.ts,.js,.json,.css,.scss,.html}\"",
-    "format:all": "prettier --config ./.prettierrc.json --write \"src/{app,environments,assets}/**/*{.ts,.js,.json,.css,.scss,.html}\"",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "format:all": "prettier --config ./.prettierrc.json --write \"src/{app,environments,assets}/**/*{.ts,.js,.json,.css,.scss,.html}\""
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.0-next.0",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^8.2.0-next.0",
-    "@angular/compiler": "^8.2.0-next.0",
-    "@angular/core": "^8.2.0-next.0",
-    "@angular/forms": "^8.2.0-next.0",
-    "@angular/http": "^6.1.10",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^8.2.0-next.0",
-    "@angular/platform-browser-dynamic": "^8.2.0-next.0",
-    "@angular/router": "^8.2.0-next.0",
-    "@fortawesome/fontawesome-free": "~5.6.3",
-    "@types/chart.js": "^2.7.40",
-    "angular": "^1.7.8",
-    "angular-bootstrap-md": "^7.5.4",
-    "angular-sanitize": "^1.7.8",
-    "bootstrap": "^4.3.1",
-    "chart.js": "^2.7.2",
-    "chartist": "^0.11.0",
-    "core-js": "^2.6.5",
-    "hammerjs": "^2.0.8",
-    "ng-quill": "^4.5.0",
-    "ngx-quill-editor": "^2.2.2",
-    "npm": "^6.9.0",
-    "osmosis": "^1.1.10",
-    "perfect-scrollbar": "^1.4.0",
-    "quill": "^1.3.6",
-    "rxjs": "6.5.2",
-    "zone.js": "^0.9.1"
+    "@angular/animations": "~11.0.5",
+    "@angular/cdk": "^11.0.3",
+    "@angular/common": "~11.0.5",
+    "@angular/compiler": "~11.0.5",
+    "@angular/core": "~11.0.5",
+    "@angular/forms": "~11.0.5",
+    "@angular/material": "^11.0.3",
+    "@angular/platform-browser": "~11.0.5",
+    "@angular/platform-browser-dynamic": "~11.0.5",
+    "@angular/router": "~11.0.5",
+    "@types/chart.js": "^2.9.29",
+    "angular-bootstrap-md": "^10.1.1",
+    "bootstrap": "^4.5.3",
+    "chart.js": "^2.9.4",
+    "chartist": "^0.11.4",
+    "perfect-scrollbar": "^1.5.0",
+    "pretty-quick": "^3.1.0",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.801.0",
-    "@angular/cli": "^8.1.0",
-    "@angular/compiler-cli": "^8.2.0-next.0",
-    "@angular/language-service": "^8.2.0-next.0",
-    "@types/jasmine": "^2.8.16",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "lint-staged": "^8.2.1",
-    "ng-lint-staged": "^0.1.6",
-    "npm-run-all": "^4.1.5",
+    "@angular-devkit/build-angular": "~0.1100.5",
+    "@angular/cli": "~11.0.5",
+    "@angular/compiler-cli": "~11.0.5",
+    "@types/node": "^12.11.1",
+    "@types/jasmine": "~3.6.0",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.1.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
     "prettier": "^1.18.2",
-    "pretty-quick": "^1.11.1",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~3.4.5"
+    "typescript": "~4.0.2"
   }
 }
