@@ -1,5 +1,4 @@
-import { Asset } from './Asset';
-
+import { Asset } from './asset';
 export class Profit {
   asset: Asset[];
   percentage: number;
