@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { User } from 'src/app/account/_models/user';
+import { User } from 'src/app/account/models/user';
 import { environment } from 'src/environments/environment';
 import { Observable } from 'rxjs';
 
