@@ -5,30 +5,6 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NavigationBarComponent } from './navigation-bar/navigation-bar.component';
 import { LayoutModule } from '@angular/cdk/layout';
 import { RouterModule } from '@angular/router';
-import {
-  MatToolbarModule,
-  MatButtonModule,
-  MatSidenavModule,
-  MatIconModule,
-  MatListModule,
-  MatGridListModule,
-  MatCardModule,
-  MatMenuModule,
-  MatFormFieldModule,
-  MatInputModule,
-  MatSnackBarModule,
-  MatProgressSpinnerModule,
-  MatSelectModule,
-  MatSliderModule,
-  MatTableModule,
-  MatTooltipModule,
-  MatRadioModule,
-  MatBadgeModule,
-  MatDialogModule,
-  MatStepperModule,
-  MatSlideToggleModule,
-  MatTabsModule,
-} from '@angular/material';
 import { MainDashboardComponent } from './main-dashboard/main-dashboard.component';
 import { SettingsComponent } from './settings/settings.component';
 import { LoginComponent } from './account/login/login.component';
@@ -36,7 +12,7 @@ import { RegisterComponent } from './account/register/register.component';
 import { routing } from './app.routing';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
-import { AlertComponent } from './account/_components/alert.component';
+import { AlertComponent } from './account/components/alert.component';
 import { AboutComponent } from './about/about.component';
 import { CreateBotComponent } from './create-bot/create-bot.component';
 import { TableBotsComponent } from './table-bots/table-bots.component';
@@ -52,6 +28,28 @@ import { StrategyDialogComponent } from './strategies/strategy-dialog/strategy-d
 import { MDBBootstrapModule } from 'angular-bootstrap-md';
 import { UserRatingComponent } from './rating/user-rating/user-rating.component';
 import { StrategiesRatingComponent } from './rating/strategies-rating/strategies-rating.component';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatButtonModule } from '@angular/material/button';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatGridListModule } from '@angular/material/grid-list';
+import { MatCardModule } from '@angular/material/card';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatSelectModule } from '@angular/material/select';
+import { MatSliderModule } from '@angular/material/slider';
+import { MatTableModule } from '@angular/material/table';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatBadgeModule } from '@angular/material/badge';
+import { MatStepperModule } from '@angular/material/stepper';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatTabsModule } from '@angular/material/tabs';
+import { MatRadioModule } from '@angular/material/radio';
+import { MatDialogModule } from '@angular/material/dialog';
 
 @NgModule({
   declarations: [
@@ -84,9 +82,8 @@ import { StrategiesRatingComponent } from './rating/strategies-rating/strategies
     HttpClientModule,
     LayoutModule,
     RouterModule,
-    routing,
     FormsModule,
-    MatDialogModule,
+    routing,
 
     // MATERIAL
     MatToolbarModule,
@@ -110,8 +107,9 @@ import { StrategiesRatingComponent } from './rating/strategies-rating/strategies
     MatBadgeModule,
     MatStepperModule,
     MatSlideToggleModule,
-    MDBBootstrapModule.forRoot(),
     MatTabsModule,
+    MatDialogModule,
+    MDBBootstrapModule.forRoot(),
   ],
   entryComponents: [AddModeyToAccountComponent, BotStatsDialogComponent, StrategyDialogComponent],
   providers: [],
