@@ -1,16 +1,14 @@
-import { Component, OnInit, DoCheck } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { CreateBotService } from './create-bot.service';
 import { FormBuilder, FormGroup, Validators, FormControl } from '@angular/forms';
-import { tap } from 'rxjs/operators';
-import { MatSnackBar, MatSelectChange } from '@angular/material';
-import { AuthenticationService } from '../account/_services/authentication.service';
-import { User } from '../account/_models/user';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatSelectChange } from '@angular/material/select';
+import { AuthenticationService } from '../account/services/authentication.service';
+import { User } from '../account/models/user';
 import { Strategy } from '../models/strategy';
-import { TradingBotsService } from '../services/trading-bots.service';
-import { CreatedTradingBot } from '../Models/trading-bot-model';
-import { FinancialInstrument } from '../models/financial-instrument-enum';
-import { Industry } from '../Models/industry-enum';
-import { Asset } from '../Models/asset';
+import { CreatedTradingBot } from '../models/trading-bot-model';
+import { Asset } from '../models/asset';
+import { FinancialInstrument } from '../models/enums';
 
 @Component({
   selector: 'app-create-bot',
@@ -45,14 +43,13 @@ export class CreateBotComponent implements OnInit {
   FU = '';
   optimisation = 1;
 
-  private formGroup: FormGroup;
-  private industryControl: FormControl = new FormControl();
-  private instrumentControl: FormControl = new FormControl();
-  private defaultProfitRisk: FormControl = new FormControl('true');
+  formGroup: FormGroup;
+  defaultProfitRisk: FormControl = new FormControl('true');
+  industryControl: FormControl = new FormControl();
+  instrumentControl: FormControl = new FormControl();
 
   constructor(
     private service: CreateBotService,
-    private tradingBotService: TradingBotsService,
     private formBuilder: FormBuilder,
     private readonly snackBar: MatSnackBar,
     private authenticationService: AuthenticationService
@@ -98,8 +95,6 @@ export class CreateBotComponent implements OnInit {
   }
 
   selectionChangeInstrument(event: MatSelectChange) {
-    const value = event.value;
-
     switch (this.instrumentControl.value) {
       case FinancialInstrument[1]:
         this.service.getAssets(FinancialInstrument.Forex, 0).subscribe(res => {
@@ -126,14 +121,5 @@ export class CreateBotComponent implements OnInit {
 
   private showMessage(msg: any) {
     this.snackBar.open(msg, undefined, { duration: 2000 });
-  }
-  private formatLabel(value: number | null) {
-    if (!value) {
-      return 0;
-    }
-    if (value <= 100) {
-      return value + '%';
-    }
-    return value;
   }
 }
