@@ -1,8 +1,8 @@
 import { tap } from 'rxjs/operators';
 import { ChangeDetectorRef, Component, Input } from '@angular/core';
-import { MatSnackBar } from '@angular/material';
 import { TradingBotsService } from 'src/app/services/trading-bots.service';
 import { TradingBot } from 'src/app/models/trading-bot-model';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-bot-name',
